+++ src/components/MessageInputSmall.js
@@ -21,6 +21,8 @@ export class MessageInputSmall extends PureComponent {
     focus: PropTypes.bool.isRequired,
     /** Grow the textarea while you're typing */
     grow: PropTypes.bool.isRequired,
+    /** Specify the max amount of rows the textarea is able to grow */
+    maxRows: PropTypes.number.isRequired,
     /** Make the textarea disabled */
     disabled: PropTypes.bool,
     /** @see See [MessageInput](https://getstream.github.io/stream-chat-react/#messageinput) for doc */
@@ -77,6 +79,12 @@ export class MessageInputSmall extends PureComponent {
     maxNumberOfFiles: PropTypes.object,
     /** @see See [channel context](https://getstream.github.io/stream-chat-react/#channel) doc */
     acceptedFiles: PropTypes.object,
+    /**
+     * Custom UI component for send button.
+     *
+     * Defaults to and accepts same props as: [SendButton](https://getstream.github.io/stream-chat-react/#sendbutton)
+     * */
+    SendButton: PropTypes.oneOfType([PropTypes.node, PropTypes.func]),
   };
 
   renderUploads = () => (
@@ -131,7 +139,7 @@ export class MessageInputSmall extends PureComponent {
   render() {
     const SendButton = this.props.SendButton;
     return (
-      <div style={{ position: 'relative', zIndex: 0, width: '100%' }}>
+      <div className="str-chat__small-message-input__wrapper">
         <ImageDropzone
           accept={this.props.acceptedFiles}
           multiple={this.props.multipleUploads}
@@ -160,6 +168,7 @@ export class MessageInputSmall extends PureComponent {
                 onChange={this.props.handleChange}
                 value={this.props.text}
                 rows={1}
+                maxRows={this.props.maxRows}
                 onSelectItem={this.props.onSelectItem}
                 placeholder="Type your message"
                 onPaste={this.props.onPaste}